+++ src/types/ability.ts
@@ -1,6 +1,6 @@
-import { Entity } from '../stages/intake'
+import { ConvoState, MessageData, SlotData, SlotStatus, AbilityStatus, ValidateResult } from './state'
+import { SetSlotDataObj } from './function'
 
-// Ability and Slot 
 /**
  * Defines conversation abilities, used to control overall flow
  * that Wolf references.
@@ -50,60 +50,4 @@ export interface ShouldRunCompleteResult {
   shouldComplete: boolean,
   reason?: string,
   nextAbility?: string
-}
-
-// Get and Set Functions
-
-export interface SetSlotDataObj {
-  setSlotValue: (abilityName: string, slotName: string, value: any, runOnFill?: boolean) => void,
-  setSlotEnabled: (abilityName: string, slotName: string, isEnabled: boolean) => void
-}
-
-// State
-
-/**
- * Conversation state managed by Botbuilder
- */
-export interface ConvoState {
-  [key: string]: any
-}
-
-/**
- * Wolf's state object that facilitates management of state sytem.
- * 
- * _User should not touch this object._
- */
-export interface WolfState {
-  messageData: MessageData,
-  slotStatus: SlotStatus[],
-  slotData: SlotData[],
-  abilityStatus: AbilityStatus[]
-}
-
-export interface MessageData {
-  rawText: string,
-  intent: string | null,
-  entities: Entity[]
-}
-
-export interface SlotStatus {
-  abilityName: string,
-  slotName: string,
-  isEnabled: boolean
-}
-
-export interface SlotData {
-  abilityName: string,
-  slotName: string,
-  value: any
-}
-
-export interface AbilityStatus {
-  abilityName: string,
-  isCompleted: boolean
-}
-
-export interface ValidateResult {
-  isValid: boolean,
-  reason: string | null
 }
