+++ src/types/function.ts
@@ -1,0 +1,6 @@
+// Get and Set Functions
+
+export interface SetSlotDataObj {
+  setSlotValue: (abilityName: string, slotName: string, value: any, runOnFill?: boolean) => void,
+  setSlotEnabled: (abilityName: string, slotName: string, isEnabled: boolean) => void
+}
